requirements.txt
pyspark==3.5.0
apache-airflow-providers-apache-spark==4.9.0
apache-airflow-providers-google==10.10.1
pandas==2.1.4
requests==2.31.0